// src/interfaces/settings.ts
export interface IAppSettings {
  layout?: string;
  boxed?: boolean;
  sidebarOpened?: boolean;
  sidebarBg?: string;
  sidebarBg2?: string;
  sidebarColor?: string;
  sidebarContrast?: string;
  sidebarAccentColor?: string;
  sidebarAccentContrastColor?: string;
  topbarBg?: string;
  topbarColor?: string;
}